==> TILDA/4_Diseases/6_joint_disorders.Rmd <==
---
title: "Joint disorders"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 
<!-- # Installation, loading packages -->
```{r setup, include=FALSE}
source("../../setup.r")
```
<!-- ########################################################## --> 
<!-- # Loading DS description -->
```{r main, child='../../_DS_Rmd/joint_disordersDS.Rmd'}
```
```{r global, echo=F}
source('../../_DS_Rmd/joint_disordersDS.R')
```




# Data process



## Ireland


### Wave 1


#### Study-specific variable description



| **Name** | `ph301_03` |`ph305`|
|-|-|-|
| **Label** | `Has a doctor ever told you that you have any of the following conditions?`|`Approximately how old were you when you were first told by a doctor that you .`|
| **Table name**  | `TILDA-AMF-v16`| `TILDA-AMF-v16`|
| **Categories** | `1 =  Arthritis (including osteoarthritis, or rheumatism)`<br/>`0 = NOT Arthritis (including osteoarthritis, or rheumatism)` |`49 = Less than 50`<br/>`59 = 50-59`<br/>`69 = 60-69`<br/>`70 = 70+`|
| **Missings** |  | `-98 = Don't Know`<br/>`-1 = Not applicable` |
| **Description** |  |  |

```{r assign w1, echo=F}

opal.assign.table.tibble(o, 'TILDA_w1','TILDA.TILDA-AMF-v16',variables=list('ph301_03','ph305'), missings = T)
TILDA_w1 <- opal.execute(o,'TILDA_w1')

```

```{r local w1, echo=F}

vbl <- TILDA_w1$ph301_03
kable(Categorical_summary(vbl, missing_values = NA)[2], caption = names(TILDA_w1)[2])
kable(Categorical_summary(vbl, missing_values = NA)[1], caption = "Category")
ggplot(TILDA_w1, aes(x=factor(ph301_03))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

vbl <- TILDA_w1$ph305
kable(Categorical_summary(vbl, missing_values = c(-1,-98))[3], caption = names(TILDA_w1)[3])
kable(Categorical_summary(vbl, missing_values = c(-1,-98))[1], caption = "Category")
ggplot(TILDA_w1, aes(x=factor(ph305))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

```


#### Harmonisation algorithm

To compute `h_joint_disorders` from the study-specific variable it has to be recoded as follows:

* `ph301_03 == 0  into 0`
* `ph301_03 == 1 into 1`

To compute `h_joint_disorders_age` from the study-specific variable it has to be recoded as follows:

* `if h_joint_disorders == 1 and ph305 ==49, h_joint_disorders_age = 40`
* `if h_joint_disorders == 1 and ph305 ==59, h_joint_disorders_age = 50`
* `if h_joint_disorders == 1 and ph305 ==69, h_joint_disorders_age = 60`
* `if h_joint_disorders == 1 and ph305 ==70, h_joint_disorders_age = 70`
* `if h_joint_disorders == 1 and ph305 ==-98, h_joint_disorders_age = 998`
* `if h_joint_disorders == 0 (<==> ph305 == -1), h_joint_disorders_age = 996`
* `otherwise into 999`

**R script:**

```{r harmo w1}

# Categorical variable harmonization
TILDA_ds_w1 <- tibble(id=TILDA_w1$id)
TILDA_ds_w1$h_joint_disorders <-TILDA_w1$ph301_03

# Continuous variable harmonization
arthritis_age <- car::recode(TILDA_w1$ph305[which(TILDA_ds_w1$h_joint_disorders==1)],"49=40;59=50;69=60;-98=998")
TILDA_ds_w1$h_joint_disorders_age <- car::recode(TILDA_ds_w1$h_joint_disorders, "0=996; 1=arthritis_age")

```


#### Statistics of the new harmonized variable

```{r descript w1, echo=F}

vbl <- TILDA_ds_w1$h_joint_disorders
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[1], caption = "Category")
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[2], caption = "Type of missing")
ggplot(TILDA_ds_w1, aes(x=factor(h_joint_disorders))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label) + ylab("Frequency")

vbl <- TILDA_ds_w1$h_joint_disorders_age
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = ds_label_age)
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(TILDA_ds_w1, aes(x=factor(h_joint_disorders_age))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label_age) + ylab("Frequency")

```


#### Validation

```{r crosstabulation w1, echo=F}

continuous <- c(sum(TILDA_ds_w1$h_joint_disorders_age==996),sum(TILDA_ds_w1$h_joint_disorders_age!=996))
dichotomous <- table(TILDA_ds_w1$h_joint_disorders, useNA = "ifany")
C <- rbind(continuous,dichotomous)
colnames(C) <- c("0<-->996","1<-->(possibly missings) age values")
kable(C,caption = "Continuous vs Dichotomous")

```



### Wave 2


#### Study-specific variable description



| **Name** | `ph301_03` |`age`|`yintw (w1)`|`yintw (w2)`|
|-|-|-|-|-|
| **Label** | `ph301 - [Since our last interview] has a doctor ever told [you/Rname] that [you/`|`Age`|`Year of interview`|`Year of interview`|
| **Table name**  | `TILDA-AMF-v23`| `age`|`yintw`|`yintw`|
| **Categories** |  `1 =  Arthritis (including osteoarthritis, or rheumatism)`<br/>`0 = NOT Arthritis (including osteoarthritis, or rheumatism)`  |`continuous`|`continuous`|`continuous`|
| **Missings** |  |  |  |  |
| **Description** |  |  |  |  |

```{r assign w2, echo=F}

opal.assign.table.tibble(o, 'TILDA_w2','TILDA.TILDA-AMF-v23',variables=list('ph301_03'), missings = T)
TILDA_w2 <- opal.execute(o,'TILDA_w2')
load(paste0(datafolder,"w2/age.RData"))
TILDA_w2 <- left_join(TILDA_w2,age,by="id")
rm(age)
load(paste0(datafolder,"w1/yintw.RData"))
TILDA_w2 <- left_join(TILDA_w2,yintw, by="id")
names(TILDA_w2)[4] <- "yintw_w1"
rm(yintw)
load(paste0(datafolder,"w2/yintw.RData"))
TILDA_w2 <- left_join(TILDA_w2,yintw, by="id")
names(TILDA_w2)[5] <- "yintw_w2"
rm(yintw)

```

```{r local w2, echo=F}

vbl <- TILDA_w2$ph301_03
kable(Categorical_summary(vbl, missing_values = NA)[2], caption = names(TILDA_w2)[2])
kable(Categorical_summary(vbl, missing_values = NA)[1], caption = "Category")
ggplot(TILDA_w2, aes(x=factor(ph301_03))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

vbl <- TILDA_w2$age
kable(Continuous_summary(var = vbl, missing_values = NA)[3], caption = names(TILDA_w2)[3])
pander(Continuous_summary(vbl, missing_values = NA)$summary, caption = "Summary")
ggplot(TILDA_w2 %>% filter(!vbl %in% miss_values_vector), aes(age)) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

vbl <- TILDA_w2$yintw_w1
kable(Categorical_summary(vbl, missing_values = NA)[2], caption = names(TILDA_w2)[4])
kable(Categorical_summary(vbl, missing_values = NA)[1], caption = "Category")
ggplot(TILDA_w2, aes(x=factor(yintw_w1))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

vbl <- TILDA_w2$yintw_w2
kable(Categorical_summary(vbl, missing_values = NA)[2], caption = names(TILDA_w2)[5])
kable(Categorical_summary(vbl, missing_values = NA)[1], caption = "Category")
ggplot(TILDA_w2, aes(x=factor(yintw_w2))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

```


#### Harmonisation algorithm

To compute `h_joint_disorders` from the study-specific variable it has to be recoded as follows:

* `ph301_03 (w1) == 0 and ph301_03 (w2) == 0 into 0`
* `ph301_03 (w1) == 1 or ph301_03 (w2) == 1 into 1`
* `otherwise into 999`

To compute `h_joint_disorders_age` from the study-specific variable it has to be recoded as follows:

* `if h_joint_disorders == 1 in w2 and 0 in w1, h_joint_disorders_age = age - (year of interview (w2) - year of interview (w1))/2`
* `if h_joint_disorders == 1 in w2 and 1 in w1, h_joint_disorders_age = age from w1`
* `if h_joint_disorders == 0 (<==> ph305 == -1), h_joint_disorders_age = 996`
* `otherwise into 999`
**R script:**

```{r harmo w2}

TILDA_w2 <- left_join(TILDA_w2,TILDA_ds_w1, by = "id")

# Categorical variable harmonization
TILDA_ds_w2 <- tibble(id=TILDA_w2$id)
TILDA_ds_w2$h_joint_disorders <- rep(999,length(TILDA_w2$id))
TILDA_ds_w2$h_joint_disorders[which(TILDA_w2$ph301_03==0 & (TILDA_w2$h_joint_disorders != 1 | is.na(TILDA_w2$h_joint_disorders)))] <- 0
TILDA_ds_w2$h_joint_disorders[which(TILDA_w2$ph301_03==1 | (!is.na(TILDA_w2$ph301_03) & TILDA_w2$h_joint_disorders == 1))] <- 1

# Continuous variable harmonization
TILDA_ds_w2$h_joint_disorders_age <- rep(999,length(TILDA_w2$id))
TILDA_ds_w2$h_joint_disorders_age[which(TILDA_ds_w2$h_joint_disorders == 1 & !is.na(TILDA_w2$h_joint_disorders_age) & !TILDA_w2$h_joint_disorders_age %in% miss_values_vector)] <- TILDA_w2$h_joint_disorders_age[which(TILDA_ds_w2$h_joint_disorders == 1 & !is.na(TILDA_w2$h_joint_disorders_age) & !TILDA_w2$h_joint_disorders_age %in% miss_values_vector)]
TILDA_ds_w2$h_joint_disorders_age[which(TILDA_ds_w2$h_joint_disorders == 1 & TILDA_w2$h_joint_disorders == 0)] <- floor(TILDA_w2$age[which(TILDA_ds_w2$h_joint_disorders == 1 & TILDA_w2$h_joint_disorders == 0)] - (TILDA_w2$yintw_w2[which(TILDA_ds_w2$h_joint_disorders == 1 & TILDA_w2$h_joint_disorders == 0)] - TILDA_w2$yintw_w1[which(TILDA_ds_w2$h_joint_disorders == 1 & TILDA_w2$h_joint_disorders == 0)])/2)

```


#### Statistics of the new harmonized variable

```{r descript w2, echo=F}

vbl <- TILDA_ds_w2$h_joint_disorders
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[1], caption = "Category")
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[2], caption = "Type of missing")
ggplot(TILDA_ds_w2, aes(x=factor(h_joint_disorders))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label) + ylab("Frequency")

vbl <- TILDA_ds_w2$h_joint_disorders_age
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = ds_label_age)
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(TILDA_ds_w2 %>% filter(!vbl %in% miss_values_vector), aes(h_joint_disorders_age)) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label_age) + ylab("Frequency")

```


#### Validation

```{r crosstabulation w2, echo=F}

AfterH <- table(TILDA_ds_w2$h_joint_disorders)
X <- c(
  sum(TILDA_w2$ph301_03==0 & (TILDA_w2$h_joint_disorders != 1 | is.na(TILDA_w2$h_joint_disorders))),
  sum(TILDA_w2$ph301_03==1 | (!is.na(TILDA_w2$ph301_03) & TILDA_w2$h_joint_disorders == 1))
)
BeforeH <- c(X[1:2])
C <- rbind(continuous,dichotomous)
colnames(C) <- c("0->0","1->1")
kable(C)

```





<!--AUXILIAR FUNCTIONS-->



```{r helpfunctions, echo=F}


source("../../helpfunctions.r")

```



```{r populations-waves, echo=F}


# All study waves and populations with abbreviated and descriptive names

tilda.cw <- list(w1 = c("w1","W1"), w2 = c("w2","W2"))



```


<!--Labelling and saving-->

```{r, echo=FALSE}


# Consider only waves with some variable harmonised

l.hds <- lapply(tilda.cw, function(wname) if(exists(paste0("TILDA_ds_",wname[1]))){wname = list(get(paste0("TILDA_ds_",wname[1])),wname[2])})
nullw <- which(sapply(l.hds, is.null))

if(length(nullw)!=0){
  l.hds <- l.hds[-nullw]
}

m.hds <- list()

lmlist <- list(lapply(l.hds,function(x)x[[1]]),m.hds)

if(length(l.hds)>0){
  # Labelling categorical variable of tibbles
  lmlist <- labelling(lmlist[[1]],lmlist[[2]],hd_vbl,ds_label)
  # Labelling continuous variable of tibbles
  lmlist <- labelling_c(lmlist[[1]],lmlist[[2]],hd_age_vbl,ds_label_age)
  # Saving tibbles
  savingRD(lmlist[[1]],vbl_name=hd_vbl)
}

tildalist <- lmlist # TILDA list




```



<!--Descriptives-->

## Descriptive statistics of the harmonised variable `h_joint_disorders` accross populations and waves

Two tables are generated: 

1. Percentages of categories in each harmonised variable.

2. Frequencies and percentages of individuals with different trajectories.


```{r summ, echo=F}


# Consider only harmonised waves
l.hds <- lapply(tilda.cw, function(wname) if(hd_vbl %in% names(get0(paste0("TILDA_ds_",wname[1])))){wname})

nullw <- which(sapply(l.hds, is.null))

if(length(nullw)!=0){
  l.hds <- l.hds[-nullw]
}



if(length(l.hds)>0){
  lmlist[[1]] <- tildalist[[1]][sapply(l.hds,function(x)x[[1]])]
  lmlist[[2]] <- tildalist[[2]][sapply(l.hds,function(x)x[[1]])]
  
  # Printing table of categories
  print(knitr::kable(summaries(lmlist[[1]],lmlist[[2]],lapply(l.hds,function(x)x[[2]]),hd_vbl),caption=ds_label))
  # Printing table of trajectories
  f.dbb <- trajectories(lmlist[[2]],vbl_name=hd_vbl)
  print(knitr::kable(f.dbb))

}



```






## Summary of descriptive statistics of the harmonised variable `h_joint_disorders_age` accross populations and waves






```{r summ_c, echo=F}


# Consider only harmonised waves
l.hds <- lapply(tilda.cw, function(wname) if(hd_age_vbl %in% names(get0(paste0("TILDA_ds_",wname[1])))){wname})

nullw <- which(sapply(l.hds, is.null))

if(length(nullw)!=0){
  l.hds <- l.hds[-nullw]
}



if(length(l.hds)>0){
  lmlist[[1]] <- tildalist[[1]][sapply(l.hds,function(x)x[[1]])]
  lmlist[[2]] <- tildalist[[2]][sapply(l.hds,function(x)x[[1]])]
  
  # Printing table of categories
  print(knitr::kable(summaries_c(lmlist[[1]],lmlist[[2]],lapply(l.hds,function(x)x[[2]]),hd_age_vbl),caption=ds_label_age))
}





```



# Quality estimation

* To compute age of diagnostic in the second wave we use the harmonized variable age, which is identical with the study-specific variable age. We have to note that in this variable, the individuals with ages less than 52 and greater or equal than 82 are collapsed to 51 and 82 years old respectively.





<!-- ########################################################## -->
<!--- # Close OPAL R Session -->
```{r closeRsession, echo=FALSE}
opal.logout(o)
```
