==> TILDA/4_Diseases/9_stroke.Rmd <==
---
title: "Stroke"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 
<!-- # Installation, loading packages -->
```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_iago.r")
```
<!-- ########################################################## --> 




# Description of DataSchema variable

The description of harmonised variable is the following:

* Short name: `stroke`
* Variable label: `Stroke`
* Variable description: `Diagnosis of stroke or cerebral vascular disease since the last wave`
* Domain: `Diseases`
* Value type: `categorical`
* Category coding:

**Code** | **Category Label**
-------- | ------------------
0        | NOT A stroke (cerebral vascular disease) 
1        | A stroke (cerebral vascular disease)

* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing

* Short name: `stroke_age`
* Variable label: `Age of each episode of stroke`
* Variable description: `Age of each episode of stroke`
* Domain: `Diseases`
* Value type: `Continuous`
* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing


<!-- ########################################################## --> 
<!-- # Global variables -->
```{r global, include=FALSE, echo=F}
miss_values_vector <- c(991, 992, 993, 994, 995, 996, 997, 998, 999)
ds_label <- "Stroke"
ds_label_age <- "Age of each episode of stroke"
cat_label <- c("No" = 0, "Yes" = 1, "Missing" = 999, "Do not know" = 998, "Refuse" = 997, "Not applicable" = 996, "Does not answer" = 995, "Not attempt not done" = 994, "Disable to measure" = 993, "Impute" = 992, "CAPI interviewer error" = 991)
cont_label <- cat_label[3:11]
datafolder <- "../RData/"
#sing_idlist <- c()
```
<!-- ########################################################## --> 



# Data process



## Ireland


### Wave 1


#### Study-specific variable description



| **Name** | `ph201_06` |`ph220`|`ybirth`|
|-|-|-|-|
| **Label** | `Has a doctor ever told you that you have any of the conditions on this card?`|`In what year was your most recent stroke?`|`year of birth`|
| **Table name**  | `TILDA-AMF-v16`| `TILDA-AMF-v16`|`ybirth`|
| **Categories** | `1 = A stroke (cerebral vascular disease)`<br/>`0 = NOT A stroke (cerebral vascular disease)` |`1 = Before 2000`<br/>`2 = 2000-2006`<br/>`3 = 2007-2010`|`continuous`|
| **Missings** |  | `-98 = Don't Know`<br/>`-1 = Not applicable` |`999 = Missing`|
| **Description** |  |  |

```{r assign w1, echo=F}

opal.assign.table.tibble(o, 'TILDA_w1','TILDA.TILDA-AMF-v16',variables=list('ph201_06','ph220'), missings = T)
TILDA_w1 <- opal.execute(o,'TILDA_w1')
load(paste0(datafolder,"w1/ybirth.RData"))
TILDA_w1 <- left_join(TILDA_w1,ybirth, by="id")
rm(ybirth)

```

```{r local w1, echo=F}

vbl <- TILDA_w1$ph201_06
kable(Categorical_summary(vbl, missing_values = NA)[2], caption = names(TILDA_w1)[2])
kable(Categorical_summary(vbl, missing_values = NA)[1], caption = "Category")
ggplot(TILDA_w1, aes(x=factor(ph201_06))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

vbl <- TILDA_w1$ph220
kable(Categorical_summary(vbl, missing_values = c(-1,-98))[3], caption = names(TILDA_w1)[3])
kable(Categorical_summary(vbl, missing_values = c(-1,-98))[1], caption = "Category")
ggplot(TILDA_w1, aes(x=factor(ph220))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

vbl <- TILDA_w1$ybirth
kable(Continuous_summary(var = vbl, missing_values = NA)[3], caption = names(TILDA_w1)[4])
pander(Continuous_summary(vbl, missing_values = NA)$summary, caption = "Summary")
ggplot(TILDA_w1, aes(ybirth)) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")+xlim(1925,1965)

```


#### Harmonisation algorithm

To compute `stroke` from the study-specific variable it has to be recoded as follows:

* `ph201_06 == 0  into 0`
* `ph201_06 == 1 into 1`

To compute `stroke_age` from the study-specific variable it has to be recoded as follows:

* `if stroke == 1 and ph220 == 1, stroke_age = 999`
* `if stroke == 1 and ph220 == -98, stroke_age = 998`
* `if stroke == 1 and ph220 ==2, stroke_age = 2000-ybirth`
* `if stroke == 1 and ph220 ==3, stroke_age = 2007-ybirth`
* `if stroke == 0 (<==> ph220 == -1), stroke_age = 996`
* `otherwise into 999`

**R script:**

```{r harmo w1}

# Categorical variable harmonization
TILDA_ds_w1 <- tibble(id=TILDA_w1$id)
TILDA_ds_w1$stroke <-TILDA_w1$ph201_06

# Continuous variable harmonization
TILDA_ds_w1$stroke_age <- rep(999,length(TILDA_w1$id))
TILDA_ds_w1$stroke_age[which(TILDA_ds_w1$stroke==0)] <- 996
TILDA_ds_w1$stroke_age[which(TILDA_ds_w1$stroke==1 & TILDA_w1$ph220==-98)] <- 998
stroke_year <- car::recode(TILDA_w1$ph220[which(TILDA_ds_w1$stroke==1 & TILDA_w1$ybirth != 999 & TILDA_w1$ph220!=-98)],"1=NA;2=2000;3=2007;-98=998")
TILDA_ds_w1$stroke_age[which(TILDA_ds_w1$stroke==1 & TILDA_w1$ybirth != 999 & TILDA_w1$ph220!=-98)] <- stroke_year - TILDA_w1$ybirth[which(TILDA_ds_w1$stroke==1 & TILDA_w1$ybirth != 999 & TILDA_w1$ph220!=-98)]
TILDA_ds_w1$stroke_age <- car::recode(TILDA_ds_w1$stroke_age, "NA=999")

```


#### Statistics of the new harmonized variable

```{r descript w1, echo=F}

vbl <- TILDA_ds_w1$stroke
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[1], caption = "Category")
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[2], caption = "Type of missing")
ggplot(TILDA_ds_w1, aes(x=factor(stroke))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label) + ylab("Frequency")

vbl <- TILDA_ds_w1$stroke_age
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = ds_label_age)
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(TILDA_ds_w1, aes(stroke_age)) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label_age) + ylab("Frequency")+xlim(35,80)

```


#### Validation

```{r crosstabulation w1, echo=F}

continuous <- c(sum(TILDA_ds_w1$stroke_age==996),sum(TILDA_ds_w1$stroke_age!=996))
dichotomous <- table(TILDA_ds_w1$stroke, useNA = "ifany")
C <- rbind(continuous,dichotomous)
colnames(C) <- c("0<-->996","1<-->(possibly missings) age values")
kable(C,caption = "Continuous vs Dichotomous")

```



### Wave 2


Not found.




## Descriptive statistics of the harmonised variable accross populations and waves

Percentages of categories in each harmonised variable:


**R script:**

```{r summ}
TILDA_ds_w1$stroke <- labelled(TILDA_ds_w1$stroke, labels = cat_label)
label(TILDA_ds_w1$stroke) <- ds_label
TILDA_ds_w1$stroke_age <- labelled(TILDA_ds_w1$stroke_age, labels = cont_label)
label(TILDA_ds_w1$stroke_age) <- ds_label_age

TILDA_m_ds_w1 <- TILDA_ds_w1
TILDA_m_ds_w1$stroke <- car::recode(TILDA_m_ds_w1$stroke, "miss_values_vector=NA")

tt <- frq(TILDA_ds_w1$stroke)[[1]][,c(1:2,4)]
tt[13,] <- c("n", "sample size", 
             length(TILDA_ds_w1$stroke)
             )
names(tt)<-c("val", "label", "w1")
kable(tt)

# dbb <- merge(TILDA_m_ds_w1[,c("id","stroke")], TILDA_m_ds_w2[,c("id","stroke")], by="id", suffixes = c(".w1",".w2"), all=T)
# names(dbb) <- c("id", "w1", "w2")
# v.dbb <- paste(dbb[,2],dbb[,3],sep="")
# f.dbb <- frq(v.dbb)[[1]][,c(1,2,4)]
# kable(f.dbb[f.dbb[,3]>=0.3,])

```



# Quality estimation

* To compute age of diagnostic in the we use the harmonized variable for year of birth, which is obtained from the study-specific variable age. We have to note that in this variable, the individuals with ages less than 52 and greater or equal than 82 are collapsed to 51 and 82 years old respectively.

<!-- ########################################################## --> 
<!-- # Save -->
```{r save, include=FALSE, echo=F}

l.TILDA <- list(w1 = TILDA_ds_w1)

for(index in seq_along(l.TILDA)){
  stroke <- l.TILDA[[index]]
  save(stroke, file = paste0(datafolder,names(l.TILDA)[index],"/stroke.RData"))
}

```
<!-- ########################################################## --> 






<!-- ########################################################## -->
<!--- # Close OPAL R Session -->
```{r closeRsession, echo=FALSE}
opal.logout(o)
```
