==> TILDA/4_Diseases/13_cvd_all.Rmd <==
---
title: "Cardiovascular diseases"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 
<!-- # Installation, loading packages -->
```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_iago.r")
```
<!-- ########################################################## --> 




# Description of DataSchema variable

The description of harmonised variable is the following:

* Short name: `cvd_all`
* Variable label: `History of all Cardiovascular diseases`
* Variable description: `The participant has been diagnosed by one or more of the following conditions: angina, stroke, myocardial infarction, heart attack, coronary heart disease, congestive heart failure, heart murmur, valve disease, cerebral vascular disease`
* Domain: `Diseases`
* Value type: `categorical`
* Category coding:

**Code** | **Category Label**
-------- | ------------------
0        | No, the participant has not been diagnosed of any of these conditions
1        | Yes, the participant has been diagnosed by any of these conditions

* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing



<!-- ########################################################## --> 
<!-- # Global variables -->
```{r global, include=FALSE, echo=F}
miss_values_vector <- c(991, 992, 993, 994, 995, 996, 997, 998, 999)
ds_label <- "History of all Cardiovascular diseases"
cat_label <- c("No" = 0, "Yes" = 1, "Missing" = 999, "Do not know" = 998, "Refuse" = 997, "Not applicable" = 996, "Does not answer" = 995, "Not attempt not done" = 994, "Disable to measure" = 993, "Impute" = 992, "CAPI interviewer error" = 991)
datafolder <- "../RData/"
#sing_idlist <- c()
```
<!-- ########################################################## --> 



# Data process



## Ireland


### Wave 1


#### Study-specific variable description



| **Name** | `h_angina` |`mi_ha`|`stroke`|`ph201_04`|`ph201_09`|`ph201_07`|
|-|-|-|-|-|-|-|
| **Label** | `History of angina`|`Myocardial infarction or heart attack`|`Stroke`|`Has a doctor ever told you that you have any of the conditions on this card?`|`Has a doctor ever told you that you have any of the conditions on this card?`|`Has a doctor ever told you that you have any of the conditions on this card?`|
| **Table name**  |`h_angina` |`mi_ha`|`stroke`| `TILDA-AMF-v16`| `TILDA-AMF-v16`| `TILDA-AMF-v16`|
| **Categories** | `0 = no history of angina`<br/>`1 = history of angina` |`0 = No`<br/>`1 = Yes`|`0 = NOT A stroke (cerebral vascular disease)`<br/>`1 = A stroke (cerebral vascular disease)`|`0 = NOT Congestive heart failure`<br/>`1 = Congestive heart failure`|`0 = NOT A heart murmur`<br/>`1 = A heart murmur`|`0 = NOT Ministroke or TIA`<br/>`1 = Ministroke or TIA`|
| **Description** |  |  |  |  |  |  |

```{r assign w1, echo=F}

opal.assign.table.tibble(o, 'TILDA_w1','TILDA.TILDA-AMF-v16',variables=list('ph201_04','ph201_09','ph201_07'), missings = T)
TILDA_w1 <- opal.execute(o,'TILDA_w1')
load(paste0(datafolder,"w1/h_angina.RData"))
TILDA_w1$h_angina <- h_angina$h_angina
rm(h_angina)
load(paste0(datafolder,"w1/mi_ha.RData"))
TILDA_w1$mi_ha <- mi_ha$mi_ha
rm(mi_ha)
load(paste0(datafolder,"w1/stroke.RData"))
TILDA_w1$stroke <- stroke$stroke
rm(stroke)

```

```{r local w1, echo=F}

vbl <- TILDA_w1$ph201_04
kable(Categorical_summary(vbl, missing_values = NA)[2], caption = names(TILDA_w1)[2])
kable(Categorical_summary(vbl, missing_values = NA)[1], caption = "Category")
ggplot(TILDA_w1, aes(x=factor(ph201_04))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

vbl <- TILDA_w1$ph201_09
kable(Categorical_summary(vbl, missing_values = NA)[2], caption = names(TILDA_w1)[4])
kable(Categorical_summary(vbl, missing_values = NA)[1], caption = "Category")
ggplot(TILDA_w1, aes(x=factor(ph201_09))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

vbl <- TILDA_w1$ph201_07
kable(Categorical_summary(vbl, missing_values = NA)[2], caption = names(TILDA_w1)[3])
kable(Categorical_summary(vbl, missing_values = NA)[1], caption = "Category")
ggplot(TILDA_w1, aes(x=factor(ph201_07))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

vbl <- TILDA_w1$h_angina
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[3], caption = names(TILDA_w1)[5])
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[1], caption = "Category")
ggplot(TILDA_w1, aes(x=factor(h_angina))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

vbl <- TILDA_w1$mi_ha
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[3], caption = names(TILDA_w1)[6])
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[1], caption = "Category")
ggplot(TILDA_w1, aes(x=factor(mi_ha))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

vbl <- TILDA_w1$stroke
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[3], caption = names(TILDA_w1)[7])
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[1], caption = "Category")
ggplot(TILDA_w1, aes(x=factor(stroke))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

```


#### Harmonisation algorithm

To compute `cvd_all` from the study-specific variable it has to be recoded as follows:

* `all 0 into 0`
* `some 1 into 1`

**R script:**

```{r harmo w1}

# Categorical variable harmonization
TILDA_ds_w1 <- tibble(id=TILDA_w1$id)
TILDA_ds_w1$cvd_all <-rep(999,length(TILDA_w1$id))
TILDA_ds_w1$cvd_all[which(TILDA_w1$ph201_04==0 & TILDA_w1$ph201_09==0 & TILDA_w1$ph201_07==0 & TILDA_w1$h_angina==0 & TILDA_w1$mi_ha==0 & TILDA_w1$stroke==0)] <- 0
TILDA_ds_w1$cvd_all[which(TILDA_w1$ph201_04==1 | TILDA_w1$ph201_09==1 | TILDA_w1$ph201_07==1 | TILDA_w1$h_angina==1 | TILDA_w1$mi_ha==1 | TILDA_w1$stroke==1)] <- 1

```


#### Statistics of the new harmonized variable

```{r descript w1, echo=F}

vbl <- TILDA_ds_w1$cvd_all
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[1], caption = "Category")
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[2], caption = "Type of missing")
ggplot(TILDA_ds_w1, aes(x=factor(cvd_all))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label) + ylab("Frequency")

```


#### Validation

```{r crosstabulation w1, echo=F}

AfterH <- table(TILDA_ds_w1$cvd_all, useNA = "ifany")
X <- c(
  sum(TILDA_w1$ph201_04==0 & TILDA_w1$ph201_09==0 & TILDA_w1$ph201_07==0 & TILDA_w1$h_angina==0 & TILDA_w1$mi_ha==0 & TILDA_w1$stroke==0,na.rm=T),
  sum(TILDA_w1$ph201_04==1 | TILDA_w1$ph201_09==1 | TILDA_w1$ph201_07==1 | TILDA_w1$h_angina==1 | TILDA_w1$mi_ha==1 | TILDA_w1$stroke==1,na.rm=T)
)
BeforeH <- c(X[1:2])
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("all 0->0","some 1->1")
kable(C)

```



### Wave 2


Not found.




## Descriptive statistics of the harmonised variable accross populations and waves

Percentages of categories in each harmonised variable:


**R script:**

```{r summ}
TILDA_ds_w1$cvd_all <- labelled(TILDA_ds_w1$cvd_all, labels = cat_label)
label(TILDA_ds_w1$cvd_all) <- ds_label

TILDA_m_ds_w1 <- TILDA_ds_w1
TILDA_m_ds_w1$cvd_all <- car::recode(TILDA_m_ds_w1$cvd_all, "miss_values_vector=NA")

tt <- frq(TILDA_ds_w1$cvd_all)[[1]][,c(1:2,4)]
tt[13,] <- c("n", "sample size", 
             length(TILDA_ds_w1$cvd_all)
             )
names(tt)<-c("val", "label", "w1")
kable(tt)

# dbb <- merge(TILDA_m_ds_w1[,c("id","cvd_all")], TILDA_m_ds_w2[,c("id","cvd_all")], by="id", suffixes = c(".w1",".w2"), all=T)
# names(dbb) <- c("id", "w1", "w2")
# v.dbb <- paste(dbb[,2],dbb[,3],sep="")
# f.dbb <- frq(v.dbb)[[1]][,c(1,2,4)]
# kable(f.dbb[f.dbb[,3]>=0.3,])

```



# Quality estimation

The harmonized variable contains information about: angina, myocardial infarction, stroke, congetsive heart failure, ministroke or TIA and heart murmur.


<!-- ########################################################## --> 
<!-- # Save -->
```{r save, include=FALSE, echo=F}

l.TILDA <- list(w1 = TILDA_ds_w1)

for(index in seq_along(l.TILDA)){
  cvd_all <- l.TILDA[[index]]
  save(cvd_all, file = paste0(datafolder,names(l.TILDA)[index],"/cvd_all.RData"))
}

```
<!-- ########################################################## --> 






<!-- ########################################################## -->
<!--- # Close OPAL R Session -->
```{r closeRsession, echo=FALSE}
opal.logout(o)
```
