==> TILDA/4_Diseases/15_oral_problems.Rmd <==
---
title: "Oral Health Problems"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 
<!-- # Installation, loading packages -->
```{r setup, include=FALSE}
source("../../setup.r")
```
<!-- ########################################################## --> 
<!-- # Loading DS description -->
```{r main, child='../../_DS_Rmd/oral_problemsDS.Rmd'}
```
```{r global, echo=F}
source('../../_DS_Rmd/oral_problemsDS.R')
```





# Data process

## Ireland

### Wave 1 

#### Study-specific variable description

| **Name** | `ph507` |
|-|-|
| **Label** | `Which best describes the teeth you have?` |
| **Table name** | `TILDA-AMF-v16` |
| **Categories** | `1 = I have all my own natural teeth - none missing`<br/>`2 = I have my own teeth, no dentures - but some missing`<br/>`3 = I have dentures as well as some of my own teeth`<br/>`4 = I have full dentures`<br/>`5 = I have no teeth or dentures`|
| **Missings** | `99 = Refused`<br/> `98 = Don't know`|
| **Description** |  |

```{r assign w1, echo=F}
opal.assign.table.tibble(o, 'TILDA_w1','TILDA.TILDA-AMF-v16', variables=list('ph507'), missings = TRUE)
TILDA_w1 <- opal.execute(o,'TILDA_w1')
```

```{r local w1, echo=F}
vbl <- TILDA_w1$ph507
vbl_miss <- c(98,99)
kable(Categorical_summary(vbl, missing_values = vbl_miss)[3], caption = names(TILDA_w1)[2]) 
kable(Categorical_summary(vbl, missing_values = vbl_miss)[1], caption = "Category")
ggplot(TILDA_w1, aes(x=factor(ph507))) + geom_histogram(stat="count", fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

```

#### Harmonisation algorithm

To compute `oral_problems` from the study-specific variable it has to be recoded as follows:

* `1,2,3 into 0`
* `4,5 into 1`
* `99 into 997`
* `98 into 998`

**R script:**

```{r harmo w1}

TILDA_ds_w1 <- tibble(id=TILDA_w1$id)
TILDA_ds_w1$oral_problems <- car::recode(TILDA_w1$ph507, "c(1,2,3) = 0; c(4,5) = 1; 98 = 998; 99 = 997 ")

```

#### Statistical description of the new harmonised variable
```{r descript w1, echo=F}
vbl <- TILDA_ds_w1$oral_problems
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[1], caption = "Category")
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[2], caption = "Type of missing")
ggplot(TILDA_ds_w1, aes(x=factor(oral_problems))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label) + ylab("Frequency")
```

#### Validation
```{r crosstabulation w1, echo=F}

```







### Wave 2

Not found.






```{r helpfunctions, echo=F}


labelling <- function(l.hds,m.hds){
  
  # Labelling of the tibbles with categorical data and creating new tibbles with all missings recodified as NA

  for(name in names(l.hds)) {
    # In the aux_object we copy the old tibble to recodify all missing values as NA.
    aux_object <- l.hds[[name]]
    # Labelling of variables
    label(l.hds[[name]][[2]]) <- label(aux_object[[2]]) <- ds_label
    # Labelling of categories (for continues variables, only missing values)
    l.hds[[name]][[2]] <- labelled(l.hds[[name]][[2]], labels = cat_label)
    aux_object[[2]] <- car::recode(aux_object[[2]], "miss_values_vector=NA")
    # Labelling of categories (for categorical variables, only non-missing values)
    aux_object[[2]] <- labelled(aux_object[[2]], labels = cat_label[1:2])
  # Saving the recodified tibble in list m.hds
    m.hds[[name]] <- aux_object
    rm(aux_object)
  }
  return(list(l.hds,m.hds))

}

# Creation of summary tables for categorical data

summaries <- function(l.hds,m.hds,lnames){

  # Creation of columns with categories and labels
  t.hds <- frq(l.hds[[1]][2])[[1]][,c(1,2)] 
  # For each wave/population in l.hds, add the correponding values
  for (i in seq_along(l.hds)){
    t.hds[2+i] <- frq(l.hds[[i]][2])[[1]][,4] 
  }
  # Add sample size for each wave/population
  t.hds[13,] <- c("n", "sample size", sapply(l.hds,function(wave) length(wave[[1]]))
  )
  # Add wave/population names
  names(t.hds)<-c("val", "label",lnames)
  return(t.hds)
  
}

# Creation of trajectories table for each population

trajectories <- function(m.hds,vbl_name){
  
  # First wave data
  dbb <- m.hds[[1]][,c("id",vbl_name)]
  # Merge with next waves data
  for(ind in 2:length(m.hds)){
    dbb <- merge(dbb, m.hds[[ind]][,c("id",vbl_name)], by = "id", suffixes = c("", paste0(".",names(m.hds)[ind])), all = T)
  }
  names(dbb) <- c("id", names(m.hds))
  
  # Glue individual data through all waves into trajectories
  v.dbb <- dbb[,2]
  for(ind in 2:length(m.hds)){
    v.dbb <- paste(v.dbb,dbb[,ind+1],sep="")
  }
  # Trajectories and frequencies
  f.dbb <- frq(v.dbb)[[1]][,c(1,2,4)]
  return(f.dbb)
  
}

# Save data tables

savingRD <- function(l.hds,vbl_name){
  
  for(index in seq_along(l.hds)){
    assign(vbl_name,l.hds[[index]])
    save(vbl_name,list = vbl_name, file = paste0(datafolder,names(l.hds)[index],"/",vbl_name,".RData"))
  }

}

```




```{r summ, echo=F}

# All study waves and populations with abbreviated and descriptive names


tilda.cw <- list(w1 = c("w1","W1"), w2 = c("w2","W2"))

# Consider only harmonised waves
l.hds <- lapply(tilda.cw, function(wname) if(exists(paste0("TILDA_ds_",wname[1]))){wname = list(get(paste0("TILDA_ds_",wname[1])),wname[2])})
nullw <- which(sapply(l.hds, is.null))

if(length(nullw)!=0){
  l.hds <- l.hds[-nullw]
}

m.hds <- list()

lmlist <- list(lapply(l.hds,function(x)x[[1]]),m.hds)

if(length(l.hds)>0){
  # Labelling tibbles
  lmlist <- labelling(lmlist[[1]],lmlist[[2]])
  # Printing table of categories
  #print(knitr::kable(summaries(lmlist[[1]],lmlist[[2]],sapply(l.hds,function(x)x[[2]])),caption=ds_label))
  # Printing table of trajectories
  #f.dbb <- trajectories(lmlist[[2]],vbl_name=hd_vbl)
  #print(knitr::kable(f.dbb))
  # Saving tibbles
  savingRD(lmlist[[1]],vbl_name=hd_vbl)
}



```





# Quality estimation


<!-- ########################################################## -->
<!--- # Close OPAL R Session -->
```{r closeRsession, echo=FALSE}
opal.logout(o)
```
