==> TILDA/4_Diseases/16_multimorbid.Rmd <==
---
title: "Multimorbidity"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 
<!-- # Installation, loading packages -->
```{r setup, include=FALSE}
source("../../setup.r")
```
<!-- ########################################################## --> 
<!-- # Loading DS description -->
```{r main, child='../../_DS_Rmd/multimorbidDS.Rmd'}
```
```{r global, echo=F}
source('../../_DS_Rmd/multimorbidDS.R')
```



# Data process


## Ireland

### Wave 1

#### Study-specific variable description


| **Name** |`h_diabetes`|`h_asthma`|`h_hypertension`|`h_joint_disorders`|`h_cpd`| `stroke` |`mi_ha`|`h_angina`|
|-|-|-|-|-|-|-|-|-|
| **Label** |`History of diabetes mellitus`|`History of asthma`|`History of hypertension`|`History of arthritis, rheumatism or osteoarthritis`|`History of chronic pulmonary disorders`| `Stroke` |`Myocardial infarction or heart attack`|`History of angina`|
| **Table name**  |`tilda_w1_hd`|`tilda_w1_hd`|`tilda_w1_hd`|`tilda_w1_hd`|`tilda_w1_hd`| `tilda_w1_hd` | `tilda_w1_hd` |`tilda_w1_hd`|
| **Categories** |`0 = No`<br/>`1 = Yes`|`0 = No`<br/>`1 = Yes`|`0 = No`<br/>`1 = Yes`|`0 = No`<br/>`1 = Yes`|`0 = No`<br/>`1 = Yes`| `0 = No`<br/>`1 = Yes` |`0 = No`<br/>`1 = Yes`|`0 = No`<br/>`1 = Yes`|
| **Description** |  |  |  |  |  |  |  |  |


```{r assign w1, echo=F}

opal.assign.table.tibble(o, 'TILDA_w1','TILDA_HD.tilda_w1_hd',variables=list('h_diabetes','h_asthma','h_hypertension','h_joint_disorders','h_cpd','stroke','mi_ha','h_angina'), missings = T)
TILDA_w1 <- opal.execute(o,'TILDA_w1')

```


```{r local w1, echo=F}

vbl <- TILDA_w1$h_diabetes
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[3], caption = "h_diabetes")
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[1], caption = "Category")
ggplot(TILDA_w1, aes(x=factor(vbl))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

vbl <- TILDA_w1$h_asthma
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[3], caption = "h_asthma")
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[1], caption = "Category")
ggplot(TILDA_w1, aes(x=factor(vbl))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

vbl <- TILDA_w1$h_hypertension
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[3], caption = "h_hypertension")
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[1], caption = "Category")
ggplot(TILDA_w1, aes(x=factor(vbl))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

vbl <- TILDA_w1$h_joint_disorders
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[3], caption = "h_joint_disorders")
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[1], caption = "Category")
ggplot(TILDA_w1, aes(x=factor(vbl))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

vbl <- TILDA_w1$h_cpd
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[3], caption = "h_cpd")
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[1], caption = "Category")
ggplot(TILDA_w1, aes(x=factor(vbl))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

vbl <- TILDA_w1$stroke
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[3], caption = "stroke")
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[1], caption = "Category")
ggplot(TILDA_w1, aes(x=factor(vbl))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

vbl <- TILDA_w1$mi_ha
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[3], caption = "mi_ha")
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[1], caption = "Category")
ggplot(TILDA_w1, aes(x=factor(vbl))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

vbl <- TILDA_w1$h_angina
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[3], caption = "h_angina")
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[1], caption = "Category")
ggplot(TILDA_w1, aes(x=factor(vbl))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

```

#### Harmonisation algorithm

To compute `multimorbid` from the study-specific variables they have to be recoded as follows:

* `Multimorbid_score = sum(h_diabetes, h_asthma, h_hypertension, h_angina, mi_ha, h_joint_disorders, stroke, h_cpd)`
* `Multimorbid_status = 1 if multimorbid_score > = 2`
* `Multimorbid_status = 0 if multimorbid_score < 2`
* `otherwise into 999`

**R script:**

```{r harmo w1}

TILDA_ds_w1 <- TILDA_w1 %>% 
  mutate(dead = purrr::pmap(mutate_at(., vars(-id), `==`, 996) %>% select(., -id), all) %>% unlist()) %>%
  mutate_at(., vars(-id), ~replace(., list = which(. %in% 991:999), values = NA)) %>% 
  mutate(multimorbid_score = rowSums(select(., -1, -dead), na.rm = T)) %>%
  mutate(cardNA = rowSums(!is.na(select(., -1, -dead, -multimorbid_score)))) %>%
  mutate(cardNA = cardNA/cardNA) %>%
  mutate(multimorbid_score = multimorbid_score * cardNA) %>%
  mutate(multimorbid = case_when(multimorbid_score >= 2 ~ 1, multimorbid_score < 2 ~ 0, dead ~ 996, TRUE ~ 999)) %>%
  select(id, multimorbid)

```


#### Statistics of the new harmonized variable

```{r descript w1, echo=F}

vbl <- TILDA_ds_w1$multimorbid
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[1], caption = "Category")
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[2], caption = "Type of missing")
ggplot(TILDA_ds_w1, aes(x=factor(vbl))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label) + ylab("Frequency")

```


#### Validation

```{r crosstabulation w1, echo=F}
```




### Wave 2

#### Study-specific variable description


| **Name** |`h_hypertension`|`h_joint_disorders`|
|-|-|-|
| **Label** |`History of hypertension`|`History of arthritis, rheumatism or osteoarthritis`|
| **Table name**  |`tilda_w2_hd`|`tilda_w2_hd`|
| **Categories** |`0 = No`<br/>`1 = Yes`|`0 = No`<br/>`1 = Yes`|
| **Description** |  |  |


```{r assign w2, echo=F}

opal.assign.table.tibble(o, 'TILDA_w2','TILDA_HD.tilda_w2_hd',variables=list('h_hypertension','h_joint_disorders'), missings = T)
TILDA_w2 <- opal.execute(o,'TILDA_w2')

```


```{r local w2, echo=F}

vbl <- TILDA_w2$h_hypertension
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[3], caption = "h_hypertension")
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[1], caption = "Category")
ggplot(TILDA_w2, aes(x=factor(vbl))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

vbl <- TILDA_w2$h_joint_disorders
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[3], caption = "h_joint_disorders")
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[1], caption = "Category")
ggplot(TILDA_w2, aes(x=factor(vbl))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

```

#### Harmonisation algorithm

To compute `multimorbid` from the study-specific variables they have to be recoded as follows:

* `Multimorbid_score = sum(h_hypertension, h_angina)`
* `Multimorbid_status = 1 if multimorbid_score > = 2`
* `Multimorbid_status = 0 if multimorbid_score < 2`
* `otherwise into 999`

**R script:**

```{r harmo w2}

TILDA_ds_w2 <- TILDA_w2 %>% 
  mutate(dead = purrr::pmap(mutate_at(., vars(-id), `==`, 996) %>% select(., -id), all) %>% unlist()) %>%
  mutate_at(., vars(-id), ~replace(., list = which(. %in% 991:999), values = NA)) %>% 
  mutate(multimorbid_score = rowSums(select(., -1, -dead), na.rm = T)) %>%
  mutate(cardNA = rowSums(!is.na(select(., -1, -dead, -multimorbid_score)))) %>%
  mutate(cardNA = cardNA/cardNA) %>%
  mutate(multimorbid_score = multimorbid_score * cardNA) %>%
  mutate(multimorbid = case_when(multimorbid_score >= 2 ~ 1, multimorbid_score < 2 ~ 0, dead ~ 996, TRUE ~ 999)) %>%
  select(id, multimorbid)

```


#### Statistics of the new harmonized variable

```{r descript w2, echo=F}

vbl <- TILDA_ds_w2$multimorbid
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[1], caption = "Category")
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[2], caption = "Type of missing")
ggplot(TILDA_ds_w2, aes(x=factor(vbl))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label) + ylab("Frequency")

```


#### Validation

```{r crosstabulation w2, echo=F}
```




## Summary of descriptive statistics of the harmonised variable accross populations and waves

Two tables are generated: 

1. Percentages of categories in each harmonised variable.

2. Frequencies and percentages of individuals with different trajectories.


```{r helpfunctions, echo=F}

source("../../helpfunctions.r")

```




```{r summ, echo=F}

# All study waves and populations with abbreviated and descriptive names


tilda.cw <- list(w1 = c("w1","W1"), w2 = c("w2","W2"))

# Consider only harmonised waves
l.hds <- lapply(tilda.cw, function(wname) if(exists(paste0("TILDA_ds_",wname[1]))){wname = list(get(paste0("TILDA_ds_",wname[1])),wname[2])})
nullw <- which(sapply(l.hds, is.null))

if(length(nullw)!=0){
  l.hds <- l.hds[-nullw]
}

m.hds <- list()

lmlist <- list(lapply(l.hds,function(x)x[[1]]),m.hds)

if(length(l.hds)>0){
  # Labelling tibbles
  lmlist <- labelling(lmlist[[1]],lmlist[[2]])
  # Printing table of categories
  print(knitr::kable(summaries(lmlist[[1]],lmlist[[2]],sapply(l.hds,function(x)x[[2]])),caption=ds_label))
  # Printing table of trajectories
  f.dbb <- trajectories(lmlist[[2]],vbl_name=hd_vbl)
  print(knitr::kable(f.dbb))
  # Saving tibbles
  savingRD(lmlist[[1]],vbl_name=hd_vbl)
}



```



# Quality estimation


```{r closeRsession, echo=FALSE} 
opal.logout(o)
```
